[build,cameraserver] Make cameraserver into a vendordep
This requires manual initialization by the robot program, but removes cscore
and OpenCV from the wpilib dependency tree.
