+++ CMakeLists.txt
@@ -287,11 +287,9 @@ if(WPILIB_WITH_WPILIB OR WPILIB_WITH_SIMULATION_MODULES)
 endif()
 
 if(WPILIB_WITH_CSCORE)
-    set(CAMERASERVER_DEP_REPLACE "find_dependency(cameraserver)")
     set(CSCORE_DEP_REPLACE "find_dependency(cscore)")
     find_package(OpenCV REQUIRED)
     add_subdirectory(cscore)
-    add_subdirectory(cameraserver)
 endif()
 
 if(WPILIB_WITH_WPILIB)
@@ -300,6 +298,9 @@ if(WPILIB_WITH_WPILIB)
     set(COMMAND_DEP_REPLACE "find_dependency(commandsv2)")
     add_subdirectory(apriltag)
     add_subdirectory(wpilibc)
+    if(WPILIB_WITH_CSCORE)
+        add_subdirectory(cameraserver)
+    endif()
     add_subdirectory(commandsv2)
     add_subdirectory(romiVendordep)
     add_subdirectory(xrpVendordep)
