+++ README-CMake.md
@@ -43,7 +43,7 @@ The following build options are available:
 * `BUILD_SHARED_LIBS` (ON Default)
   * This option will cause CMake to build static libraries instead of shared libraries.
 * `WPILIB_WITH_CSCORE` (ON Default)
-  * This option will cause cscore to be built. Turning this off will implicitly disable cameraserver. If this is off, the OpenCV build requirement is removed.
+  * This option will cause cscore to be built. CameraServer is built when both this option and `WPILIB_WITH_WPILIB` are enabled. If this is off, the OpenCV build requirement is removed.
 * `WPILIB_WITH_EXAMPLES` (OFF Default)
   * This option will build C++ examples.
 * `WPILIB_WITH_GUI` (OFF Default)
@@ -118,9 +118,10 @@ cmake_minimum_required(VERSION 3.11)
 project(vision_app) # Project Name Here
 
 find_package(wpilib REQUIRED)
+find_package(cameraserver REQUIRED)
 
 add_executable(my_vision_app main.cpp) # executable name as first parameter
-target_link_libraries(my_vision_app cameraserver ntcore cscore wpiutil)
+target_link_libraries(my_vision_app cameraserver)
 ```
 
 If you want to use other libraries or are building a robot program, `wpilibc` and `hal` should be added in the `target_link_libraries` function, along with any other libraries you plan on using, e.g. `wpimath`.
@@ -135,7 +136,7 @@ After that, run `cmake --build .`. That will create your executable. Then you sh
 
 ## Using vendordeps
 
-Vendordeps are not included as part of the `wpilib` CMake package. However, if you want to use a vendordep, you need to use `find_package(VENDORDEP)`, where `VENDORDEP` is the name of the vendordep (case-sensitive), like `xrpVendordep` or `romiVendordep`. Note that commandsv2, while a vendordep in normal robot projects, is not built as a vendordep in CMake, and is instead included as part of the `wpilib` CMake package. After you used `find_package`, you can reference the vendordep library like normal (using `target_link_libraries`).
+Vendordeps are not included as part of the `wpilib` CMake package. However, if you want to use a vendordep, you need to use `find_package(VENDORDEP)`, where `VENDORDEP` is the name of the vendordep (case-sensitive), like `cameraserver`, `xrpVendordep`, or `romiVendordep`. Note that commandsv2, while a vendordep in normal robot projects, is not built as a vendordep in CMake, and is instead included as part of the `wpilib` CMake package. After you used `find_package`, you can reference the vendordep library like normal (using `target_link_libraries`).
 
 ## Troubleshooting
 Below are some common issues that are run into when building.
