+++ cameraserver/cameraserver-config.cmake.in
@@ -4,6 +4,7 @@ include(CMakeFindDependencyMacro)
 @DATALOG_DEP_REPLACE@
 @NTCORE_DEP_REPLACE@
 @CSCORE_DEP_REPLACE@
+@WPILIBC_DEP_REPLACE@
 find_dependency(OpenCV)
 
 @FILENAME_DEP_REPLACE@
