+++ commandsv2/BUILD.bazel
@@ -113,7 +113,6 @@ wpilib_java_library(
     plugins = ["//javacPlugin:plugin"],
     visibility = ["//visibility:public"],
     deps = [
-        "//cscore:cscore-java",
         "//hal:hal-java",
         "//ntcore:ntcore-java",
         "//wpiannotations",
