+++ commandsv3/BUILD.bazel
@@ -54,7 +54,6 @@ wpilib_java_library(
     plugins = ["//javacPlugin:plugin"],
     visibility = ["//visibility:public"],
     deps = [
-        "//cscore:cscore-java",
         "//hal:hal-java",
         "//ntcore:ntcore-java",
         "//wpiannotations",
