+++ docs/build.gradle
@@ -156,8 +156,30 @@ configurations {
     }
 }
 
+def javaDocProjects = [
+    project(':apriltag'),
+    project(':cameraserver'),
+    project(':commandsv3'),
+    project(':cscore'),
+    project(':epilogue-runtime'),
+    project(':hal'),
+    project(':ntcore'),
+    project(':wpiannotations'),
+    project(':commandsv2'),
+    project(':wpilibj'),
+    project(':wpimath'),
+    project(':wpinet'),
+    project(':wpiunits'),
+    project(':wpiutil'),
+    project(':romiVendordep'),
+    project(':xrpVendordep'),
+    project(':datalog'),
+]
+
 task generateJavaDocs(type: Javadoc) {
-    classpath += project(":wpilibj").sourceSets.main.compileClasspath
+    javaDocProjects.each {
+        classpath += it.sourceSets.main.compileClasspath
+    }
     options.links("https://docs.oracle.com/en/java/javase/25/docs/api/")
     // workaround for opencv site blocking javadoc tool. If the link is changed,
     // docs/opencv/element-list must be redownloaded
@@ -193,23 +215,9 @@ task generateJavaDocs(type: Javadoc) {
     options.addBooleanOption('html5', true)
     options.linkSource(true)
     dependsOn project(':wpilibj').generateJavaVersion
-    source project(':apriltag').sourceSets.main.java
-    source project(':cameraserver').sourceSets.main.java
-    source project(':commandsv3').sourceSets.main.java
-    source project(':cscore').sourceSets.main.java
-    source project(':epilogue-runtime').sourceSets.main.java
-    source project(':hal').sourceSets.main.java
-    source project(':ntcore').sourceSets.main.java
-    source project(':wpiannotations').sourceSets.main.java
-    source project(':commandsv2').sourceSets.main.java
-    source project(':wpilibj').sourceSets.main.java
-    source project(':wpimath').sourceSets.main.java
-    source project(':wpinet').sourceSets.main.java
-    source project(':wpiunits').sourceSets.main.java
-    source project(':wpiutil').sourceSets.main.java
-    source project(':romiVendordep').sourceSets.main.java
-    source project(':xrpVendordep').sourceSets.main.java
-    source project(':datalog').sourceSets.main.java
+    javaDocProjects.each {
+        source it.sourceSets.main.java
+    }
     source configurations.javaSource.collect { zipTree(it) }
     include '**/*.java'
     failOnError = true
