+++ wpilib-config.cmake.in
@@ -6,8 +6,6 @@ find_dependency(Threads)
 @EIGEN_SYSTEM_REPLACE@
 
 @APRILTAG_DEP_REPLACE@
-@CAMERASERVER_DEP_REPLACE@
-@CSCORE_DEP_REPLACE@
 @DATALOG_DEP_REPLACE@
 @HAL_DEP_REPLACE@
 @NTCORE_DEP_REPLACE@
