+++ wpilibc/src/test/native/c/test.c
@@ -2,6 +2,5 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-#include "wpi/cs/cscore_c.h"
 #include "wpi/hal/HAL.h"
 #include "wpi/nt/ntcore.h"
