+++ wpilibc/wpilibc-config.cmake.in
@@ -4,8 +4,6 @@ include(CMakeFindDependencyMacro)
 @DATALOG_DEP_REPLACE@
 @WPIMATH_DEP_REPLACE@
 @NTCORE_DEP_REPLACE@
-@CSCORE_DEP_REPLACE@
-@CAMERASERVER_DEP_REPLACE@
 @HAL_DEP_REPLACE@
 
 @FILENAME_DEP_REPLACE@
