+++ wpilibcExamples/src/main/cpp/snippets/HttpCamera/cpp/Robot.cpp
@@ -19,6 +19,7 @@
 class Robot : public wpi::TimedRobot {
  public:
   Robot() {
+    wpi::CameraServer::Initialize();
     // We need to run our vision program in a separate thread. If not, our robot
     // program will not run.
     std::thread visionThread(VisionThread);
