+++ xrpVendordep/xrpvendordep-config.cmake.in
@@ -2,8 +2,6 @@ include(CMakeFindDependencyMacro)
 @WPIUTIL_DEP_REPLACE@
 @DATALOG_DEP_REPLACE@
 @NTCORE_DEP_REPLACE@
-@CSCORE_DEP_REPLACE@
-@CAMERASERVER_DEP_REPLACE@
 @HAL_DEP_REPLACE@
 @WPILIBC_DEP_REPLACE@
 @WPIMATH_DEP_REPLACE@
